+++ protocols/gossipsub/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Update to `libp2p-swarm` `v0.38.0`.
 
+- Update to `libp2p-core` `v0.35.0`.
+
 # 0.39.0
 
 - Update to `libp2p-core` `v0.34.0`.