+++ transports/uds/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.34.0 [unreleased]
+
+- Update to `libp2p-core` `v0.35.0`.
+
 # 0.33.0
 
 - Update dependencies.